+++ src/main/java/com/team9ookie/dangdo/service/FileService.java
@@ -1,6 +1,6 @@
 package com.team9ookie.dangdo.service;
 
-import com.team9ookie.dangdo.dto.FileDto;
+import com.team9ookie.dangdo.dto.file.FileDto;
 import com.team9ookie.dangdo.repository.FileRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Service;
 @AllArgsConstructor
 public class FileService {
 
-    private FileRepository fileRepository;
+    private final FileRepository fileRepository;
 
     public void saveFile(FileDto fileDto) {
         fileRepository.save(fileDto.toEntity());
